requirements.txt
Django==3.2.3
djangorestframework==3.12.4
requests==2.26.0
drf-spectacular==0.18.0